move option helpers to own file
